app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
selenium
undetected-chromedriver
webdriver-manager
pandas
openpyxl
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
dnspython
email-validator
user-agent
aiofiles
gunicorn